core/requirements.txt
-i https://pypi.org/simple
appdirs==1.4.4
argcomplete==3.0.8 ; python_full_version >= '3.6.0'
argparse==1.4.0
asgiref==3.6.0 ; python_version >= '3.7'
asttokens==2.2.1
backcall==0.2.0
click==8.1.3 ; python_version >= '3.7'
colorama==0.4.6 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
coverage==7.2.3
cssbeautifier==1.14.7
cx-freeze==6.14.9
cx-logging==3.1.0 ; sys_platform == 'win32'
decorator==5.1.1 ; python_version >= '3.5'
django==4.2
django-browser-reload==1.8.0 ; python_version >= '3.7'
django-extensions==3.2.1
django-mathfilters==1.0.0
django-pytest==0.2.0
django-tailwind==3.5.0
django-widget-tweaks==1.4.12
djhtml==3.0.6
djlint==1.25.0
editorconfig==0.12.3
exceptiongroup==1.1.1 ; python_version < '3.11'
executing==1.2.0
html-tag-names==0.1.2 ; python_version >= '3.7' and python_version < '4.0'
html-void-elements==0.1.0 ; python_version >= '3.7' and python_version < '4.0'
idna==3.4
iniconfig==2.0.0 ; python_version >= '3.7'
ipython==8.12.0
isort==5.12.0
jedi==0.18.2 ; python_full_version >= '3.6.0'
jsbeautifier==1.14.7
lief==0.13.0 ; sys_platform == 'win32'
matplotlib-inline==0.1.6 ; python_version >= '3.5'
packaging==23.1 ; python_version >= '3.7'
parso==0.8.3 ; python_full_version >= '3.6.0'
pathspec==0.11.1 ; python_version >= '3.7'
pickleshare==0.7.5
pillow==9.5.0
pluggy==1.0.0 ; python_full_version >= '3.6.0'
prompt-toolkit==3.0.38 ; python_version >= '3.7'
pure-eval==0.2.2
pygments==2.15.1 ; python_version >= '3.7'
pypng==0.20220715.0
pyserial==3.5
pytest==7.3.1 ; python_version >= '3.7'
pytest-django==4.5.2
python-escpos==2.2.0
pyusb==1.2.1 ; python_full_version >= '3.6.0'
pywin32==306
pyyaml==6.0 ; python_full_version >= '3.6.0'
qrcode==7.4.2 ; python_version >= '3.7'
regex==2023.3.23 ; python_full_version >= '3.8.0'
setuptools==67.7.2 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4 ; python_version >= '3.5'
stack-data==0.6.2
tomli==2.0.1 ; python_version < '3.11'
tqdm==4.65.0 ; python_version >= '3.7'
traitlets==5.9.0 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
tzdata==2023.3
wcwidth==0.2.6
win32printing==0.1.3
